+++ app.go
@@ -8,7 +8,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/adrg/xdg"
 	"github.com/pkg/errors"
 	"github.com/vovan-ve/sple-desktop/internal/backend"
 	"github.com/vovan-ve/sple-desktop/internal/config"
@@ -37,9 +36,9 @@ func (a *App) startup(ctx context.Context) {
 	// Perform your setup here
 	a.ctx = ctx
 
-	configDir := filepath.Join(xdg.ConfigHome, "sple")
-	if err := config.EnsureConfigDir(configDir); err != nil {
-		runtime.LogErrorf(ctx, "config dir: %v", err)
+	configDir := config.GetConfigsDir()
+	if err := config.EnsureDir(configDir, "config dir"); err != nil {
+		runtime.LogErrorf(ctx, "%v", err)
 		return
 	}
 
@@ -274,17 +273,21 @@ func (a *App) SaveFileAs(blob64 helpers.Blob64, baseFileName string) {
 		return
 	}
 
-	filepath, err := runtime.SaveFileDialog(a.ctx, runtime.SaveDialogOptions{
+	path, err := runtime.SaveFileDialog(a.ctx, runtime.SaveDialogOptions{
 		DefaultFilename: baseFileName,
 		Title:           "Save As",
 	})
-	if err != nil || filepath == "" {
+	if err != nil || path == "" {
 		return
 	}
-	err = files.NewFile(filepath).Write(b)
+	err = files.NewFile(path).Write(b)
 	return
 }
 
 func (a *App) SetIsDirty(isDirty bool) {
 	a.isDirty = isDirty
+}
+
+func (a *App) GetAppInfo() string {
+	return config.ReportInfo()
 }