+++ frontend/wailsjs/go/main/App.d.ts
@@ -4,6 +4,8 @@ import { files } from '../models';
 
 export function CreateFile(arg1:string,arg2:string):Promise<string>;
 
+export function GetAppInfo():Promise<string>;
+
 export function OpenFile(arg1:boolean):Promise<Array<files.WebFileRef>>;
 
 export function SaveFileAs(arg1:string,arg2:string):Promise<void>;