+++ internal/logging/log_skip.go
@@ -1,0 +1,9 @@
+//go:build !production
+
+package logging
+
+import (
+	"github.com/wailsapp/wails/v2/pkg/logger"
+)
+
+func GetLogger() logger.Logger { return nil }